LayoutTests/fast/forms/checkbox-default-value.html

<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="console"></div>
<script>
    var el = document.createElement("input");
    el.setAttribute("type", "checkbox");
    shouldBe('el.value', '"on"');
    shouldBe('el.setAttribute("value", "foo"); el.value', '"foo"');
    shouldBe('el.checked = true; el.value', '"foo"');
    shouldBe('el.removeAttribute("value"); el.value', '"on"');
    shouldBe('el.value = "foo"; el.getAttribute("value")', '"foo"');
</script>
</body>
